+++ src/euroeval/dataset_configs/icelandic.py
@@ -158,3 +158,23 @@ MULTI_WIKI_QA_IS_CONFIG = DatasetConfig(
     languages=[ICELANDIC],
     unofficial=True,
 )
+
+ICELANDIC_LANG_TESTS_CONFIG = DatasetConfig(
+    name="icelandic-lang-tests",
+    pretty_name="Icelandic Language Tests",
+    source="EuroEval/icelandic-lang-tests",
+    task=KNOW,
+    languages=[ICELANDIC],
+    unofficial=True,
+    val_split=None,
+)
+
+ICELANDIC_MATH_TESTS_CONFIG = DatasetConfig(
+    name="icelandic-math-tests",
+    pretty_name="Icelandic Mathematics Tests",
+    source="EuroEval/icelandic-math-tests",
+    task=KNOW,
+    languages=[ICELANDIC],
+    unofficial=True,
+    val_split=None,
+)